Added league to teams

+++ src/main/java/com/example/thymeleaf/repository/LeagueRepository.java
@@ -3,6 +3,6 @@ package com.example.thymeleaf.repository;
 import com.example.thymeleaf.model.entity.League;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeagueRepository extends JpaRepository<League, Long> {
-    League findLeagueById(Long id);
+public interface LeagueRepository extends JpaRepository<League, Integer> {
+    League findLeagueById(Integer id);
 }
